Simplify ToOrdinalSuffix tests
